Fix the consume bug
Fix the bug that the consume recipe functionality is unusable
